requirements-complex.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
azure-cognitiveservices-speech==1.41.1
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.2.2 
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gotrue==2.11.1
#gradio==5.0.0
#gradio_client==1.4.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.27.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
janus==1.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-core==0.3.30
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langgraph==0.2.64
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.11
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
miniaudio==1.61
msgpack==1.1.0
multidict==6.1.0
nltk==3.9.1
numpy==1.26.4
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.14
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.52
pillow==10.4.0
postgrest
propcache==0.2.1
protobuf==4.25.5
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyht==0.0.28
PyJWT==2.10.1
pyngrok==7.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
#realtime==2.0.0
redis==6.0.16
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
ruff==0.9.2
s3transfer==0.10.4
safehttpx==0.1.6
semantic-version==2.10.0
sentry-sdk==2.19.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.37
starlette==0.37.2
storage3
StrEnum==0.4.15
supabase==2.5.1
supafunc
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.21.0
tomlkit==0.12.0
tqdm==4.67.1
twilio==9.4.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
./vocode-ss
watchfiles==1.0.3
websockets==11.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
